requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.3
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.14
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.7.4
fastapi==0.95.2
fastapi-cli==0.0.5
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
langchain==0.0.251
langsmith==0.0.18
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.4
pyarrow==12.0.1
pydantic==1.10.12
pydantic_core==2.23.4
pydeck==0.8.0
pydentic==0.0.1.dev3
Pygments==2.15.1
Pympler==1.0.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.17
python-stdnum==1.20
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.19
starlette==0.27.0
streamlit==1.25.0
streamlit-chat==0.1.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
typer==0.13.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
uvicorn==0.32.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.24.0
websockets==14.1
wheel==0.44.0
yarl==1.9.2
youtube-transcript-api==0.6.2
zipp==3.16.2
